Move `CDN_URL` to urls.js utility
